Move a symbolization related functionality into symbolize module
This change moves a bunch of the symbolization related functionality
into the newly introduced symbolize module. This separation will make it
easier to conditionally compile parts of the crate in the future.
