+++ src/resolver.rs
@@ -12,10 +12,10 @@ use crate::ksym::KALLSYMS;
 use crate::log;
 use crate::maps;
 use crate::maps::Pid;
+use crate::symbolize::AddressLineInfo;
 use crate::util;
 use crate::util::uname_release;
 use crate::Addr;
-use crate::AddressLineInfo;
 use crate::FindAddrOpts;
 use crate::SymbolInfo;
 use crate::SymbolSrcCfg;
